:recycle: Splitting the SBOM creation in its own package to have a more readable code
