+++ pkg/reporter/sbom/cyclonedx_1_4.go
@@ -1,0 +1,29 @@
+package sbom
+
+import (
+	"github.com/CycloneDX/cyclonedx-go"
+	"github.com/google/osv-scanner/pkg/models"
+	"io"
+)
+
+func ToCycloneDX14Bom(_ io.Writer, packageSources []models.PackageSource) *cyclonedx.BOM {
+	bom := cyclonedx.NewBOM()
+	components := make([]cyclonedx.Component, 0)
+	bom.JSONSchema = cycloneDx14Schema
+	bom.SpecVersion = cyclonedx.SpecVersion1_4
+
+	uniquePackages := groupByPackage(packageSources)
+
+	for packageURL, packageDetail := range uniquePackages {
+		component := cyclonedx.Component{}
+		component.Name = packageDetail.Name
+		component.Version = packageDetail.Version
+		component.BOMRef = packageURL
+		component.PackageURL = packageURL
+		component.Type = componentType
+		components = append(components, component)
+	}
+	bom.Components = &components
+
+	return bom
+}
